Update script to create subjects table
